Add diagnostic handling to SonicAnemometer
The SonicAnemometer class now handles the status value
for RMY81000 sonics correctly. If the user asks for an "ldiag"
value, then u,v,w,tc will be flagged as missing if the parsed
status is not 0.

Checking of variable names for "diag" and "ldiag" is now done
for all sonics in the base class.