Change some xsd:normalizedString to xsd:string
In a normalizedString, tabs, returns and newlines are converted
to spaces. We don't want that in things like prompts and
message separators.

We generally use \r or \xHH sequences to encode those characters,
which NIDAS converts after the XML parse, so this isn't generally
an issue, but we might as well be explicit.