Fix bug in multicast connections
The (very complicated!) code keeps track of requests made
so that the multicast data packets are sent on the correct
interface.

For multicast requests, the data packets are not send back to
the requesting socket, but to a different port number.

This multicast support sure complicates things - not sure
it is worth it, but it's working now.