Don't create samples for all possible sensor types
Previously the validate method created sample ids
for all possible sensor types, for mote 0, giving them
generic variable names. Programs, such as ck_xml and
cockpit which queried the samples from all sensors
whould then include these extraneous samples.

Those unconfigured sample ids are no longer generated,
and so cockpit shouldn't create plots of those unnecessary
variables.

If an unconfigured sample is encounted in the raw sample stream,
it is ignored (skipped) for sensor types of WST_IGNORE. If
the sensor type is then not WST_NOWARN, a warning log message
is generated.

For right now, there are no WST_NOWARN sensors, but that could be
changed for a sensor that we would like to see in the output of
data_dump, for example, but don't want warnings.