consolidate sample statistics in the SampleCounter class
Refactor CounterClient to use a single map of SampleCounter instances, and
track all known sensors and sample tags.  By default it only reports on
received samples, but soon it will be able to report samples which were not
received.  Adjust loop so it can repeat when the period alarm signals.