Reworked Watlow process method
Check the first three bytes of each sample, using the length field for
the number of bytes to parse in each sample.

The process method now follows the sensor configuration in the XML
for the number of samples to expect, the number of variables in each sample,
and applies an optional conversion (e.g. linear, polynomial) to each variable.

For samples 1 and 2 the signed values are still divided by 10, but that
could be replaced by conversions in the XML if other conversions are wanted.

This is an example of a process method for a sensor that extracts
all useful fields from binary data, and parses fields based on both
what is found in the input data and what is requested in the XML, filling
the remainding requested variables with NAN.

Only every 1/1000th warning about bad data records is logged, preventing
a saturation of the system logs, which could result from a misconfigured
sensor.  This check of "!(_nwarn++ % N)" is used for logging messages
in other sensor classes.

Added a faster CRC algorithm found in the "Modicum Modbus Protocol Reference Guide", which can be found on the web. This reduced the processing time of
Watlow records by only about 12%, but I figured it was worth leaving in.
It returns the same result as the original CRC calc, after an endian tweak.

There is some code in the constructor, #ifdef'd out at this time, that
reports the CRC for the current prompts. If a prompt is changed or added,
one could tweak that code to figure out the prompt CRC.