Replace addSampleTag() with validate()
Also move code from A2DSensor::fromDOMElement() to validate().

When addSampleTag() is called for a sensor that is found in
the <sensorcatalog>, the SampleTag has only been initialized
from the catalog, and so may be incomplete. Using validate()
avoids this issue.

Remove extraneous addSampleTag() and fromDOMElement() methods
in DSC_A2DSensor.