Hack to address failures in test/serial_sensor
Every once in a while the serial sensor test fails, with the
archive files showing a few less samples than are expected.

As a hack, in a likely futile attempt to avoid looking into the
problem in more detail, expand sorter lengths, heaps and
lateSampleCacheSize settings.

Since the data is coming in through pseudo-terminals, the time-tags
that are assigned to the samples are dependent on system load.