Fix 16 bit reads on Vulcan when overloaded
If no sample is available from the buffer for sending to the
user side, data is read and discarded from the card.
The read uses the inw_16o macro, which adds ISA_16BIT_ADDR_OFFSET
to the given address. On all systems except
Vulcans, this ISA_16BIT_ADDR_OFFSET=0 and this macro is just inw.
On Vulcans ISA_16BIT_ADDR_OFFSET is non-zero. So its a but

Doing
    inw_16o(brd->addr + read_address_offset + ISA_16BIT_ADDR_OFFSET);
results in the offset being added twice to the address.

This bug would only show up on Vulcans if the CPU gets behind and
there are no more samples in the buffer. So perhaps we've never
seen it.