add opt to treat TSoil as uint16 / Fixes ISFS-174
https://jira.eol.ucar.edu/browse/ISFS-174 has more information,  but
Tsoil during the last 2 projects was getting over 32.767C, which, on an
unsigned 16 bit int, rolls over, to a negative temp.  This commit adds
the ability to modify the sensor_catalog.xml Tsoil cfg with an
additional bool paramater named "TsoilUnsigned16" which, if true, will
allows NIDAS to treat the incoming 'Tsoil' as a unsigned 16 bit int.
This allows for hot environment (0deg to 65.536C) to be properly parsed,
but at the exclusion of working below 0 degrees.

This is a hack that should be avoided for future cases, but has been
added to aid with processing PERDIGAO and ARTSE data sets.

The firmware has not modified, so there is a very good reason to assume
this will happen again.  For future growth, this should be modified to
load a signed int32 bit values rather than int16s which allos for maintaining
the same level of resolution (more actially) with minimal changes.