add some flexibility to parsing WXT messages
Allow the header field (eg 0R0, 1r0) in WXT scanf formats to match any ID
and either upper or lower case, since only distinguishes whether the
messages contain checksums, and at the moment the checksum is ignored.
Also allow format fields to be skipped if their first two characters do not
match the field in the sample, in which case the variable corresponding to
that format field is filled with nans.  This allows the same scanf format
to be used for mulitple WXT sensors (eg rooftop weather stations) even if
not all the same fields are enabled in all the sensors.