Rudimentary support for Metek Sonics
This takes original C++ code from Jakob Mann <jmsq@dtu.dk> and
the Risø-R-1658(EN) paper, and modifies it for use within nidas.
The paper does not completely match the provided code, which is
more reliable.  Jakob has been given notice and may eventually update
the paper to match the algorithm he provided.

Functionally there are 3 LUTs that contain static corrections to apply
to uvw and a linear tranposition to apply to t. The values oif the table
are linerally interpolated.

A 'metek' boolean parameter flag has been added to the sonic
configuration, which is picked up and checked in validate() and
process() checks this internal flag in order to apply the correction
before other corrections can be applied.

This assumes that the normal Metek corrections have been applied, but I
cannot find reliable evidence of the Metek settings, so the code assumes
that the standard Metek corrections have been enabled.

-Nick Potts