change DSMSensor::convertVariable() to Variable::convert()
And change the API to pass a results pointer instead of the 'limitcheck'
flag.  That gives the caller more flexibility in how the filtered values
are copied into the output values, if they need to be handled differently.
Also avoid passing the DSMSensor pointer just to check
getApplyVariableConversions(), and instead call that method on the
Variable's Site member pointer directly.