NAN out Temperature on Metek if U, V, or W are NAN
As part of the Perdigao processing code, there have arrisen some odd
values coming from a sonic where U, V are 'normal', bu W is overly
large, by 100 times.  This causes the correction code to 'correct' U, V,
W to be NAN, but leaves the temperature value as some non-sensical
number.  This quick fix cause the Metek code to set U, V, W and T to
NAN when this occurs, as the Temperature is quite tied to all three.