support kernel timer api as of 4.15
The timer callbacks now pass a single argument of struct timer_list*.  In
the old interface the device pointer was passed through an unsigned arg and
cast.  So now callbacks pass struct timer_list* through both APIs: in the
old interface the pointer is set in the data member of struct timer_list,
in the new interface there is no data member, since the callback always
passes the pointer to the timer_list struct.  The pointer to the containing
device structure is retrieved with container_of().  Once the old API does
not need to be supported, container_of can be replaced with the newer
from_timer() macro.