UNTESTED changes to build arinc driver for 4.15 kernel timer API
For pre-4.15 kernels, set the .data member to the timer_list pointer, same
as is forced with the 4.15 api.  The kernel timer callbacks are now passed
the timer_list pointer rather than a channel number.  For the
arinc_timesync callback, the callback parameter is not used anyway.  For
the arinc_sweep callback, the arinc_dev pointer containing the sweeper
timer_list is found using container_of(), and then the channel number is
derived by subtracting the base of the chn_info array.

Always call del_timer_sync(), there is no need to check the (now removed)
data member to see if a timer was active.

If the IRIG callbacks are ever used again, it might be a good idea to
register the timer_list pointer as the callback parameter (instead of
channel number), same as is done with the kernel timer callbacks, so the
callbacks are implemented the same regardless of how they are being
triggered.