New time-based average filter for 16-bit A2D data
A time-based average filter, aka timeavg, is like a simple
boxcar average, but the set of samples that are averaged
is based on the input sample time tags, rather than simply
the next N samples, as is the case with the boxcar filter.

Adding this filter also resulted in some refactoring.