add -D,--data to report data values averaged over recording period
The SampleCounter now accumulates sums of all variables in processed
samples, so it can print an average value with the variable name in the
report.  It also counts and reports the number of NaNs seen in a variable
over the period.  For raw samples, it just prints the last seen raw sensor
message, same as data_dump.