New TimetagAdjuster class for processed time tags
The time tag adjuster can be used for processed samples with a
fixed rate.  It monitors an estimated latency in the assignment
of the original time tags to each raw sample in the archive.
Corrected time tags are generated using the estimated latency and
the fixed time delta.

Currently the CharacterSensor::process() method will use the
TimetagAdjuster on processed samples if the <sample> element
has a ttadjust="N" parameter where N is the number of seconds
in the latency computation and is positive.

This has been given limited testing on HoneywellPPT samples
from the C130.