Use INT_MAX for initial tdiffminUsec
Time difference between adjusted and raw time series may be more
than 2 sample delta-Ts over the adjustment period if a sensor's
rate is significantly different than the advertised rate.

This arose when testing for a sensor whose actual sampling rate
was about 49.5 Hz instead of the advertized 50Hz.
So after an adjustment period of 10 seconds there are 495 samples
instead of 500, which creates a step jump of 5*0.02 = 0.1 seconds
every 10.  Not good.

May want to improve TimetagAdjuster to smooth the jumps, and
spread the jump over the next adjustment period. In doing
this, the adjuster would learn the actual sampling rate.