Alter Metek corrections to allow for  steep angles
The PERDIGAO data set has large periods where the wind data is mostly UP
or DOWN, which was not run as part of the Metek's original correction
routines (in those periods, the values were nailed to NAN, which marked
data as erroneous).  This just pegs values outside the correction
acceptance angle range to the min / max correction value, which was
suggested by the original author of the paper.