Merge working port config with desired
Since desired port config is updated during fromDOMElement which is
invoked in a base class, the values modified via the serialSensor
element attributes must be merged into the desired port config in the
super class.

Also updated "confirm >" logic in GILL2D to not actually search for the
confirm text, but just send the confirmation command immediately.