handle errors from database posts, report the total measurements posted
influxdb returns errors in json format from database http posts, so read
and parse the json messages.  Stash the error state in the database class,
and allow the nidas code beneath the sample receive() method to interrupt
the app with an exception generated from the database error message.  so
now data_influxdb exits on the first error, either from connection failures
to the database, or error results from the influxdb server.