config nc_server in one place, using parseconfig.ParseConfig
src/nidas/SConscript was adding the nc_server libs using a CheckConfig,
MergeFlags, Configure check, and then src/nidas/dynld/isff/SConscript was
adding the libs and cflags again.  Instead, add libs and cflags in one
place in src/nidas/SConscript, so it is not needed in isff/SConscript, and
use parseconfig.ParseConfig, since that is equivalent to CheckConfig and
MergeFlags.  I think this fixes what I broke in armbe cross-builds on
CentOS 7, and hopefully is a little cleaner and less confusing.