Fix non-increasing times in MessageStreamScanner
If time tags of samples on a message stream are not increasing
it must be due to a step change in the system clock, or
due to system buffering, where the receipt times of two
buffers are closer than should be possible at the
given baud rate.

A backwards step change is detected when the time of a buffer
read is earlier that the previous read.  In this case, warnings
are issued about non-increasing time tags, but no correction
is done.  In a bad situation, maybe the backwards step change
is huge, and correct, and it seems better to pass on the time
tags, rather than alter them, and some other method must be used
to fix the time tags.

If a non-increasing time tag is not the result of a step change,
then a warning is issued, and the time tag adjusted to the
last sample tag for the stream plus one microsecond. In this
way, a time-sort of the samples will not place them out-of-order.