Tighten up the NMEA time field parsing
Use "%n" to check that 6 characters are parsed for HHMMSS and YYMMDD,
and that the next character is a comma, or a period in the case of HHMMSS.FF.
Screen parsed time and date fields (except year) for sanity.
If any of these tests fail, all fields in the sample are set to NaN,
even those that may have been successfully parsed.  The time tag of the
NaN'd sample is set to the raw sample time.

This code is necessary because of a major problem with I2C Ublox GPS's
used by ISFS on the R Pis. There are so many bad samples, that a
troubling number get past the checksums.

Need to make sure this is not destroying good data before committing to
master.