add nextArg() method, reset file length in each output specifier
allows nidsmerge -i to consume multiple non-option arguments.  If an output
length is not specified in the output file specifier, then it resets to 0.
This avoids using the length attached to a previous output specifier, since
that seems unreasonable.  Either an output length suffix must be specified,
or the app must provide a way to set a default for all output specifiers,
like nidsmerge -l.