got one test to pass, but added some others that don't...
This branch looks close to being functional enough to merge.  It looks like
the <repeat> element needs to be fleshed out, and then at least reasonable
time schedules could be described in the xml.  At the moment this is
capable of describing when sensors come on and offline.  The next piece is
instrumenting sample processing to direct samples to the right sensor
depending on the time.  For the RELAMPAGO pressure sensor swap, a few
checks will have to be changed to allow multiple sensors with the same IDs
in the XML, as long as they do not collide in time.