fix bug in recent refactor of SampleInputStream::search()
Searching for samples across input changes in a SampleInputStream was not
working, causing the first file in a fileset to be skipped.  Filesets also
confuse the bad block tracking because the stream offset resets to zero.
So now everything that must happen on isNewInput() is consolidated in the
handleNewInput() method.  That method reports the current block stats and
resets them, and it also reads the input header.  The file name of the
input channel is now also cached in the SampleInputStream so it can be
referenced even after the input has advanced to a new file.