remove my bad idea: let users specify time filtering explicitly
For the nidas utilities which automatically enable bad sample filtering by
setting max sample length, this code would trigger a default time window of
24 hours into the future. That's fine for processing past data, but of
course it's a problem for a long-running real-time nidas process.  It was a
bad idea to assume a reasonable time window when no processing window is
already known.  We could of course extend this to be years into the future,
but really, the user should not have to use time filtering if they don't
want it.  There should not be a hidden automatic time filter, no matter how
reasonable we think it is.