map sensor samples to measurements in influx schema
Since all the variables in a nidas sample tag share the same timestamp, the
sample streams from each sensor are the point sources, the time series.
This change tries to give the measurement names a useful name, by using a
combination of sensor catalog name (ie IDREF) and the first variable name
in the sample.  However, there is no expectation that the schema is user
friendly, it is meant to optimize influx storage and access.  The metadata
and other schema info, especially variable units, now must come from
somewhere else.