Don't zero statusOR after IRIG_GET_STATUS ioctl
These changes were mostly motivated by running pc104sg on a Vortex
where for some as yet unknown reason, the interrupt service routine is
not being called.

If the ISR is not being called, then the status is not updated.
If it is zeroed after the value is returned in the ioctl, the
next status more likely will not not match the actual status.

In init method, set CLOCK_STATUS_NOSYNC and CLOCK_SYNC_NOT_OK bits
from the status register, so that initially those bits will be correct.