differentiate variant build dirs by os release as well as arch
The goal is to build for multiple targets (arch/os) from the same source
tree, without interfering with each other.  So tests need to be built as
variants also, but that doesn't work yet.  Also Configure() output still
needs to be redirected to the variant location.  Someday the variant dir
will be a scons option, so variant builds can be created for other
differences, like for developing multiple branches from one source tree.