simplify for rootless podman containers
It is ok to run as root inside the container, since that automatically maps
to the user running the container in the host, and that simplifies all the
staging and mounting.  Install libjsoncpp-dev for target arch.  Do not
install eol_scons, in favor of mounting from the developer install on the
host.