switch armbe container to centos7
this sets up like current barolo builds, except there is still a
problem with xmlrpc++ and pkg-config I think, the XmlRpc
library is not found like it is on barolo.  This also removes the
special build user setup, and just expects root in the container
to map to the user running the container.  Also expects nidas
source and eol_scons to be mounted into the container as
done by ncidas.sh script.