Read all available events from inotify fd
Use O_NONBLOCK on inotify fd, and read events until EAGAIN.

If you're watching file.txt and someone does:
    cat "99" >> file.txt
This will generate two events, IN_MODIFY and IN_CLOSE_WRITE, rather
than one event with both bits set.

After a poll detects activity on inotify fd, one or both of these
events may be available to read.  If the IN_MODIFY was consumed in the
previous poll/read, the next poll on inotify will return the
IN_CLOSE_WRITE event. In handling this event, the file is not read,
and the previous version would return exhausted=false, which the
SensorHandler would mistakenly log as a full buffer read for the sensor.