refactor update algorithm to tighten sample times
rather than replaying queued samples right before the report, reset
stats and filter samples when starting to collect samples for the
next time period.  include a lag in the collection time, reject samples
earlier than the current period, cache samples past the current period
but do not accumulate them, so that the only samples included in the
stats and json output are the ones which fall within the specified
time period.