Non-blocking I/O, only seek to end in first open
When a file is reopened after a move, or the inode has changed,
don't seek to end, instead start reading at beginning. This should
reduce the possibility of missing records on a busy log file after
a logrotate.

Use O_NONBLOCK on the file to avoid any possibility of blocking.