more bad sample filtering for GRAINEX
Since the sample time tag is first in the sample header,
and written little-endian no less, it is possible for a sample
header to overlap with the end of a bad block, so the magnitude
of the time tag can be reasonable (ie impossible to filter by a
range limit) but still off.  Therefore there is no choice but to skip
the first unfiltered sample at the end of a bad block.

This happens the other way also.  Since the sample type is in the
most significant 6 bits, it is the last to be written.  If that overlaps
with the start of a bad block, then everything except the type
can be bad.  So add the 'raw' filter rule to force the sample type
to be CHAR_ST (0).  In theory the last unfiltered sample before a
bad block should also be skipped, but that would be a significant
refactor, so see how far this much gets us.  I expect the corrupted
data will cause problems and need to be addressed.