Use do_div for 64 bit divide and mod on 32 bit sys
Trying to do 64 bit int divides or modulus in kernel modules
on a 32 bit system will result in undefined __divdi3 or
__moddi3 on module load.

tv_sec member of thiskernel_timespec_t is 64 bit on kernels > 3.17.
So on 32 bit systems one must use do_div for divide and modulus on
tv_sec.

Use int64_t for 64 bit time, uint64_t is not necessary.