Updates for running reprepro in a container
Do a test signing with gpg2.  This provides a check for version
incompatibility between gpg2/reprepro and the gpg-agent. This
incompatibiity arose when running gpg2/reprepro in a Ubuntu 18
container, talking to gpg-agent running on a CentOS7 host.

If packages already exist in the repo, remove them and try again.
This will cause apt-get users to get checksum errors when installing
the packages, so they will have to do apt-get update to get the new
package lists and checksums. This is unavoidable if we want to make
sure the packages are installed on the repo successfully.

In run_docker.sh, shutdown the gpg-agent on the host before
starting the container to avoid the gpg2/gpg-agent version incompatiblilty.