Updates for building under ubuntu bionic
Ubuntu bionic actually checks Build-Depends in debian/control.
Require xmlrpc++-dev only for i386. Tweak depends for libgsl.
Left in Build-Depends of eol-scons, though we usually clone it
into the docker images.

Mount eol_scons in the containers if it exists next to nidas.

Add lintian to the docker image. It is used in building the eol-scons
package.