TimetagAdjust mods
Separate logic for backwards time and gaps. Report counts for backwards and
gaps separately.

On a gap, don't reset the calculated _dtUsecCorr. This helps in doing
ttadjust for a "bursty" sensor, such as ARINC label 074, which is
nominally 10 sps, but 9/10 of the dt's in a second are about 0.004 sec,
followed by a dt of about 0.9 sec.  One probably should not bother trying
to ttadjust those time tags, but at least now it is able to adjust to
the high sample rate, without resetting the calculated dt every second.

On backwards time, just return input timetag and don't change any other
accumulated values.  I have not tested this on a time series with backwards
but it seems like the thing to do.

Add octalLabel argument to log() method to output ARINC labels in octal.