Rework of SyncRecordSource, tweak to SampleTracer
Pass all the LOG_CONTEXT args, e.g. __FILE__ and __FUNCTION__,
to the SampleTracer constructor so that the file and function of
the LogContext in SampleTracer is set from where the SampleTracer
was created, not, as previously, the file and function of the
SampleTracer constructor.

SyncRecordSource now just tries to poke the next sample into the next
slot in the sync record, rather than computing the slot index each time.
This method should be more tolerant of latency jitter in the time tags.

A sample is discarded if:
1. the sample timetag is earlier than the first sync record in memory
2. the sample timetag is more that NSLOT_LIMIT delta-Ts earlier than the
   time associated with the next slot.

The current value of NSLOT_LIMIT is 3.

If the sample timetag is more than NSLOT_LIMIT delta-Ts later than the
next slot time, then the slot index is recomputed, skipping over
slots, leaving nans in the skipped slots. The total of skipped slots
is kept for each sample.

If a sample is occuring faster than the configured rate, or the
samples occur in bursts, then it may fill all the slots in the second
sync record, before the first sync record is written out.  When this
happens the value of the last slot is overwritten and an overwritten counter
incremented.

In the SyncRecordSource destructor the number of discarded and
overwritten samples is reported for each sample id. Depending on
a #define, the number of skipped slots can be logged also.

Add a COMPUTE_OFFSET macro, which if defined, the time offset of the
first sample in the sync record is written into the sync record
for each variable.  If COMPUTE_OFFSET is not defined, then all offsets
will be zero. Right now it is defined.

Assemble the information necessary for inserting a sample into a sync
record in a SyncInfo class. A map of SyncInfos by sample ids replaces
a bunch of vectors.

Log the number if discarded, overwritten and, optionally, skipped samples
in the SyncRecordSource destructor.

Increase the precision of the sample rates in the sync record header.
The previous value of setprecision(2) was insufficient for rates like
3.125 Hz.