More rework and cleanup of SyncRecordSource
Fix bug in previous rework: off-by-one error in writing values
to sync record.

SyncInfo now has methods for incrementing, decrementing,
and computing the slot index - calling methods in
SyncRecordSource if needed.  Those methods can ensure
indices are valid.

New method for inserting NaNs when rates are non-integral.

The fractional second offset for each sample is now computed
from the minimum amount of time over the second that a sample
is later than its slot time, rather than the offset of the
first sample in the second. This should reduce some variablility
in the time offsets.