Never adjust a time tag later
The assumption of this algorithm is that measured time tags are only
late, never earlier than "truth". Previous version of ttadjust
was adjusting some time tags later, due to errors in estimating the
actual reporting rate.

When a production run is done, summary ttadjust statistics are logged.
The "max late" field is the maximum adjustment backwards (earlier) that
was applied to time tags for a sample. It is worth checking this
value to get an idea of how much correction was applied.

The choice of the sampleGap and adjustPeriod parameters will effect the
corrrection.  It would be nice to come up with a way to automatically
choose good values for those parameters.