Improvements to TimetagAdjuster
ttadjust is now enabled in the XML with one positive value:
        <sample ... ttadjust="1">
There is no need to guess at a value for an expected latency gap.

By default, the number of points in the adjustment analysis period is
5 for low rates and the sample rate for rates over 5.

As the minimum tdiff is computed over these points, a check is also made
to see if tdiff is decreasing, and the process is continued until tdiff
is no longer decreasing.  This method works for periods where the data
system buffers up a large number of samples, and then a number of
samples appear in the archive with closely spaced time tags.

Added a doc/timetags.md file to document the method.

The sample id is passed to the TimetagAdjuster constructor, and so
doesn't need to be passed on each call to adjust().