Improvements to SyncRecordSource
Address problem with a sensor reporting interval that is slightly fast
or slow, such that eventually a sync record slot needs to be skipped or overwritten.

If a sample time tag is too early or too late by TDIFF_CHECK_USEC
for outOfSlotMax number of samples in a row, then a slot is skipped
for late time tag or overwritten for an early time tag.

TDIFF_CHECK_USEC is currently 2000 (2 milliseconds), and outOfSlotMax is
std::max(3 * USECS_PER_SEC / dtUsec, 5), the number of samples in 3
seconds.

Also commit some edits to timetags.md