handle newinput after partial sample data read
If the end of a file (newinput) was hit before reading all of a
sample's data, then nextSample() would return 0, even if keepreading
was true.  That should not happen.  If keepreading is true, then
nextSample() needs to either return a sample or throw an exception.
Fix this by handling eof and newinput at the start of the loop, and
restarting the loop any time those events occur, for both sample
header reads and sample data reads.