Add chrony values to status output for RAF use
Added a new sensor class, ChronyLog, derived from
WatchedFileSensor, and a virtual interface, ChronyStatusNode,
with a printChronyStatus() method, which outputs a html
table row containing the NTP stratum and time offset, that
has been parsed from the chrony log messages.

The DSMServerStat run method has been expanded, creating
a list of those sensors that can be dynamic_cast'd to a
ChronyStatusNode, and periodically calls the printChronyStatus()
method on those nodes.

index.php should now grab the chrony status elements and render
the html, but this has not been tested.

I've tested the output of printChronyStatus(), and that a list
of ChronyStatusNodes can be created from a list of DSMSensor.

No changes should be necessary to status_listener.

Added some documentation on the format of the XML status packets.