use std::isnan to avoid ambiguities
something about gcc 4.8.5 on RHEL7 causes isnan to be ambiguous in the
default namespace.  other attempts to disambiguate it appear to not
be portable across all our gcc versions, so probably best to just
explicitly qualify it as std::isnan everywhere, and include cmath
as needed to get the std namespace definition.