allow computed u,v in same sample
Unlike process() in other sensors which add derived u,v, WxtSensor was
adding a new sample even if the u,v variables were in the same sample
as wspd,wdir.  Fix that, and also work around the subsequent warnings
about scanf fields, by replacing the inherited validateSscanfs() with an
explicit call to a private method at the end of WxtSensor::init().