keep copy of ident in Logger
Allocate a copy of ident in the Logger instance to guarantee the memory
is valid for the life of a syslog connection.  I think this fixes a bug
mentioned in ISFS-354, but of course it will not fix DSM3 until the
branches are merged.

Since the const char* pointer is not being passed directly to openlog(),
the constructor signature now accepts const std::string&.

Writing a test to tax the memory allocation and destroy a Logger
instance revealed an inconsistency in the Logger api.  The Logger
destructor is no longer public, it has been replaced with
destroyInstance().