Updates for moving from docker to podman
See the Docker/Podman wiki page on the EOL SEW wiki:
https://wiki.ucar.edu/pages/viewpage.action?pageId=364844003&src=contextnavpagetreemode

Leave/set the USER to root:root in our images. Then the root user inside
the container is mapped to the user running podman.  Simplifies things

Use the --group-add=keep-groups podman run option so that the
supplement group ids are enabled in the container.