Introduced initFTDI to the repo. This is a small utility intended to run at boot/reset time, to  preset the bit-bang mode and state of the FTDI interfaces.
It has been determined that the first time the mode is set on an interface, it changes the state of any pins on the interface that are designated as outputs. This has had some unintended side-effects; see Jira issue ISFS-451. Running this program at boot time forces the pins into a known/preferred state.