INTERFACE_C on device I2C is special. A value of 0xDF is consistently read from that interface after power-up/reset.  The upper nibble of pins controls the wifi/default LEDs/switches:
  bit-7 -- def LED
  bit-6 -- wifi LED
  bit-5 -- def switch
  bit-4 -- wifi switch

The upper nibble being 0xD, or 1101, should indicate that both LEDs should be on (and wifi switch "on", whatever that means). But they are not initially; after this program runs, having set the pins as output pins and rewriting the 0xDF, the LEDs do come on.

Although the switches and LEDs are not currently incorporated into the DSM behavior, leaving the pins in this state would seem to be inconsistent. Thus, this program explicitly sets the pins to off/0; e.i, it writes 0x0F to I2C/ifaceC (more correctly, it writes 0x0?  where "?" is whatever was initially read from the lower nibble).