Add XmlRpc code to A2D_Serial.
For A2D_Serial and DSC_A2DSensor, ioctl() will not be needed or used.

ncar_a2d: iotctl() to control card diagnostic voltage
Diamond: Will use D2A.  Channel 0 for A2D 0-15 and channel 1 for 16-32.
gpDAQ: Uses RS232.

We will maintain the state of the diagnostic voltage in the SensorClass,
whereas the ncar_a2d maintains that over in the driver (i.e. the need
for ioctl()).