Updates to NMEA filtering in tee_i2c
If "-G" option preceeds a pty name argument, the RMC and GGA messages
are parsed and those that don't meet expectations are not sent to that
pty.

The intent is to use this for screening NMEA bad messages received
over an I2C, when the interface suffers from the clock-stretcing issue
(as seen with ublox 5/6 on an RPi2). The filtered output can then be used
by the chrony/NTP NMEA/PPS reference clock, which *should* reduce issues
with bad clocks.

Added a testing option, whereby NMEA messages can be read and output to
text files. If the I2C address in the runstring is 0, then the i2cname
is interpreted as a text file.

This was tested on an arbitray 12 hour file from CHEESEHEAD and it
reported 10 bad RMC/GGA messages which passed the checksum test, including
5 bad times and 1 bad date:
2021-11-25,11:03:39|INFO|# total bad checksums: 1991 , # total bad fields: 10
2021-11-25,11:03:39|INFO|# bad RMC:date=1, lon=1, status=1, time=3, bad GGA:lat=2, time=2