Added start_podman to /opt/nidas/bin
Part of nidas-devel package - couldn't think of a better place
to put it.

start_podman is a script like nidas/scripts/run_podman.sh,
but since it is in /opt/nidas/bin it will likely be found on the
$PATH for nidas users. It mounts the current directory to the
container (along with .gnupg and /net/ftp/pub/archive/software/debian).
Should be useful for general debian builds.