Remove clearVoltages() from dtor, fix counts calc
ck_aout creates the sensor, sets the voltage, then destroys the sensor.
As a result the voltages were always being set back to 0V in the
destructor.  This change was added in Nov 2021 for autocal purposes.
The autocal sequence will have to call clearVoltages() itself.

Also found elementary math bug in calculation of counts, which would
only show up if the minimum voltage value was not 0V.