Updates to the new ModbusRTU sensor
It is basically working, as tested on a bench Vortex DSM, with a
DS18B20 temperature probe interfaced through a Eletechsup R46CA01
on /dev/ttyS2, reading once a second.

However, the modbus_read_registers() function fails intermittently,
roughly once every minute.  Many errors don't have a known errno value:
    Error: IOException: /dev/ttyS2: modbus_read_registers: Unknown error 112345691
Other times the error is
    Error: IOException: /dev/ttyS2: modbus_read_registers: Connection timed out

ModbusRTU has a crude workaround for these errors, in that when they occur,
the modbus_read_registers thread exits.  The SensorHandler in Nidas will
then get a timeout on reading the pipe, which results in a close and re-open
of the serial port, and data will resume, until the next error.

I have a hunch the "Unknown error" is due to a bad checksum in the data read.
My next hunch is that the errors may be related to problems with I/O on a
a half-duplex 485 device. Past experience is that this can be tricky, in that if the
transmitter is enabled, it causes glitches in the received data. This
can be addressed by doing a set/clear of RTS before/after a write.
However, with libmodbus, we cannot do anything between the underlaying
write and the read in modbus_read_registers().

I tried using modbus_rtu_set_serial_mode(,MODBUS_RTU_RS485)
but this resulted in "Inappropriate ioctl for device". It would be
good to look at the modbus code to see what ioctl is being attempted.

The CTI serial board and driver have ioctls for 485. Perhaps we should
test the temperature probe on a CTI port.