Remove legacy zeroth bin for size distribution probes.
These probes would be the SPP100, 200, 300, UHSAS, and 2D probes.  I
have commented out the #define that engages this extra bin (front
padded in the historgam.  This works in conjunction with nimbus' removal
of its use also.  And nidas XML files need to have 1 subtracted from the
length of the histogram.

TI3GER will be the first project that uses this.